packages/embeddings-creation/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
backoff==2.2.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudevents==1.9.0
colorama==0.4.6
deprecation==2.1.0
Flask==2.2.3
frozenlist==1.3.3
functions-framework==3.3.0
google-api-core==2.11.0
google-auth==2.16.2
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.2
openai==0.27.2
packaging==23.0
pandas==1.5.3
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
regex==2023.3.23
requests==2.28.2
rsa==4.9
six==1.16.0
tiktoken==0.3.1
tqdm==4.65.0
urllib3==1.26.15
watchdog==2.3.1
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
